refactor(hodo-list): Rename type files
