+++ apps/hodo-list/src/atoms/tasks-atom.ts
@@ -1,6 +1,6 @@
 import { atomWithStorage } from "jotai/utils";
 
-import type { Task } from "@/types/Task";
+import type { Task } from "@/types/task.type";
 
 export const tasksTodayAtom = atomWithStorage<Task[]>("tasks-today", []);
