+++ apps/hodo-list/src/atoms/tasks-records-atom.ts
@@ -1,5 +1,5 @@
 import { atomWithStorage } from "jotai/utils";
 
-import type { TasksRecord } from "@/types/tasks-record";
+import type { TasksRecord } from "@/types/tasks-record.type";
 
 export const tasksRecords = atomWithStorage<TasksRecord[]>("tasks-records", []);
