+++ apps/hodo-list/src/types/tasks-record.type.ts
@@ -1,4 +1,4 @@
-import type { Task } from "@/types/Task";
+import type { Task } from "@/types/task.type";
 
 export type TasksRecord = {
   date: number;
