+++ apps/hodo-list/src/utils/task-filters.ts
@@ -1,4 +1,4 @@
-import type { Task } from "@/types/Task";
+import type { Task } from "@/types/task.type";
 
 /**
  * @param tasksToday
